+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The `Unreleased` section name is replaced by the expected version of next releas
 ### Added
 
 - Streams Scheduler: Tune memory consumption re write positions of inactive streams [#94](https://github.com/jet/propulsion/pull/94) 
+- Streams Scheduler: Add `purgeInterval` to reclaim memory used by write positions of inactive streams [#95](https://github.com/jet/propulsion/pull/95) 
 
 ### Changed
 ### Removed